MDEV-21839: Handle crazy offset to SHOW BINLOG EVENTS
Problem:
=======
SHOW BINLOG EVENTS FROM <"random"-pos> caused a variety of failures as
reported in MDEV-18046. They are fixed but that approach is not future-proof
as well as is not optimal to create extra check for being constructed event
parameters.

Fix:
===
When binlog checksum is disabled perform sequential scan of the requested log
to validate the crazy offset. Starting from offset 4 read the event_length of
next_event in the binary log. Using the next_event length advance current
offset to point to next event. Repeat this process till the current offset is
less than or equal to crazy offset. If current offset is higher than crazy
offset provide appropriate invalid input offset error.